+++ frontend/next-sitemap.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  siteUrl: 'https://gov-status.vercel.app',
+  generateRobotsTxt: true,
+  sitemapSize: 5000,
+}
