+++ frontend/package.json
@@ -5,16 +5,17 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "postbuild": "next-sitemap",
     "start": "next start",
-    "lint": "eslint",
-    "deploy": "vercel --prod",
-    "preview": "vercel"
+    "lint": "eslint"
   },
   "dependencies": {
     "clsx": "^2.1.1",
     "ioredis": "^5.8.0",
     "lucide-react": "^0.544.0",
     "next": "15.5.4",
+    "next-seo": "^6.8.0",
+    "next-sitemap": "^4.2.3",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "recharts": "^3.2.1"
