+++ frontend/public/google43f45afa11f71b9d.html
@@ -1,0 +1,1 @@
+google-site-verification: google43f45afa11f71b9d.html
