+++ frontend/public/robots.txt
@@ -1,0 +1,9 @@
+# *
+User-agent: *
+Allow: /
+
+# Host
+Host: https://gov-status.vercel.app
+
+# Sitemaps
+Sitemap: https://gov-status.vercel.app/sitemap.xml
