+++ frontend/public/sitemap-0.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
+<url><loc>https://gov-status.vercel.app/contributors</loc><lastmod>2025-09-30T16:41:31.517Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gov-status.vercel.app/report</loc><lastmod>2025-09-30T16:41:31.522Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gov-status.vercel.app/analytics</loc><lastmod>2025-09-30T16:41:31.522Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gov-status.vercel.app/services</loc><lastmod>2025-09-30T16:41:31.522Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://gov-status.vercel.app</loc><lastmod>2025-09-30T16:41:31.522Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+</urlset>
