+++ frontend/public/sitemap.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+<sitemap><loc>https://gov-status.vercel.app/sitemap-0.xml</loc></sitemap>
+</sitemapindex>
