+++ frontend/src/app/api/redis/dashboard/route.ts
@@ -427,7 +427,7 @@ export async function GET() {
       agencyStats: agencyStatsArray
     }, {
       headers: {
-        'Cache-Control': 's-maxage=60, stale-while-revalidate=30'
+        'Cache-Control': 's-maxage=600, stale-while-revalidate=300'
       }
     });
   } catch (error) {
